move log tests to separate files
